Fix imports for PromptResponse being moved to ./types
