+++ js/packages/teams-ai/src/models/LlamaModel.ts
@@ -4,8 +4,9 @@ import { TurnContext } from 'botbuilder';
 import { Colorize } from '../internals/Colorize';
 import { Memory } from '../MemoryFork';
 import { Message, PromptFunctions, PromptTemplate } from '../prompts';
-import { PromptCompletionModel, PromptResponse } from '../models';
+import { PromptCompletionModel } from '../models';
 import { Tokenizer } from '../tokenizers';
+import { PromptResponse } from '../types';
 
 export interface LlamaModelOptions {
     apiKey: string;
